feat: 🎸 gpt融合知识

+++ Server/app/service/review.py
@@ -8,6 +8,20 @@ from app.models.similar_knowledge import (
 )
 from sqlalchemy.orm import Session
 from app.models.enums import ReviewStatus, ReviewType, KnowledgeItemType
+from langchain.prompts import PromptTemplate
+from langchain.llms import AzureOpenAI
+from langchain.chains import LLMChain
+from langchain.chat_models import ChatOpenAI
+from langchain.prompts.chat import (
+    ChatPromptTemplate,
+    SystemMessagePromptTemplate,
+    HumanMessagePromptTemplate,
+)
+import os
+from dotenv import load_dotenv
+import json
+
+load_dotenv()
 
 
 def query_similar_knowledge(vectors, docs):
@@ -76,12 +90,48 @@ def create_review_item(similar_knowledge: SimilarKnowledgeCreate):
     db.close()
     return entity
 
+def validate_knowledge_existed(similar_knowledge: SimilarKnowledge):
+    supabase = create_supabase_client()
+    res = supabase.table("knowledge").select("*",count="exact").eq("id", similar_knowledge.old_knowledge_id).execute()
+    return res.count > 0
 
-def fusion_knowledge():
-    pass
+def fusion_knowledge(silimar_knowledge: SimilarKnowledge):
+    old_knowledge_existed = validate_knowledge_existed(silimar_knowledge)
+    if not old_knowledge_existed:
+        return
+    
+    os.environ["OPENAI_API_TYPE"] = "azure"
+    os.environ["OPENAI_API_BASE"] = os.getenv("AZURE_OPENAI_API_BASE")
+    os.environ["OPENAI_API_KEY"] = os.getenv("AZURE_OPENAI_API_KEY")
+    os.environ["OPENAI_API_VERSION"] = "2023-05-15"
+    llm = ChatOpenAI(engine="gpt-35-turbo")
+    prompt_template = PromptTemplate.from_template(
+        "融合以下两段文字，1、```{knowledge1}```, 2、```{knowledge2}```"
+    )
+    prompt = prompt_template.format(
+        knowledge1=silimar_knowledge.new_knowledge,
+        knowledge2=silimar_knowledge.old_knowledge,
+    )
+    template = """你负责融合两段不同的知识，融合后的知识应该是一段新的知识，不需要关注融合后的知识是否有意义。你的返回应该是一段json, 包括这些字段: `knowledge: 融合后的知识`"""
+    system_message_prompt = SystemMessagePromptTemplate.from_template(template)
+    human_template = "{text}"
+    human_message_prompt = HumanMessagePromptTemplate.from_template(human_template)
+
+    chat_prompt = ChatPromptTemplate.from_messages(
+        [system_message_prompt, human_message_prompt]
+    )
+    chain = LLMChain(llm=llm, prompt=chat_prompt)
+    res = chain.run(prompt)
+    
+    silimar_knowledge.new_knowledge = json.loads(res)["knowledge"]
+    add_knowledge(silimar_knowledge)
 
 
 def replace_knowledge(silimar_knowledge: SimilarKnowledge):
+    old_knowledge_existed = validate_knowledge_existed(silimar_knowledge)
+    if not old_knowledge_existed:
+        return
+    
     add_knowledge(silimar_knowledge)
     supabase = create_supabase_client()
     supabase.table("knowledge").delete().eq(
@@ -113,7 +163,7 @@ def add_knowledge(silimar_knowledge: SimilarKnowledge):
 def update_review_item(id: int, db: Session, action: ReviewType):
     entity = db.query(SimilarKnowledge).filter(SimilarKnowledge.id == id).first()
     if action == ReviewType.FUSION.name:
-        fusion_knowledge()
+        fusion_knowledge(entity)
     elif action == ReviewType.REPLACE.name:
         replace_knowledge(entity)
     elif action == ReviewType.ADD.name:
